+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add support for Python 3.10 and 3.11
 - Configure HTTP client to use a network timeout of 10 seconds
 - Use `versioningit` for versioning
+- Tests: Remove dependency on Apiary Mock Server API
 
 ## 4.0.0 (2021-09-19)